requirements.txt
fastapi
uvicorn[standard]
openai
numpy
scikit-learn
pydantic
python-multipart
python-dotenv
